// packages/plugins/plugin-openapi/src/types/plugin.ts
/* -------------------------------------------------------------------

                   🗲 Storm Software - Powerlines

 This code was released as part of the Powerlines project. Powerlines
 is maintained by Storm Software under the Apache-2.0 license, and is
 free for commercial and private use. For more information, please visit
 our licensing page at https://stormsoftware.com/licenses/projects/powerlines.

 Website:                  https://stormsoftware.com
 Repository:               https://github.com/storm-software/powerlines
 Documentation:            https://docs.stormsoftware.com/projects/powerlines
 Contact:                  https://stormsoftware.com/contact

 SPDX-License-Identifier:  Apache-2.0

 ------------------------------------------------------------------- */

import type { OpenAPISchema } from "@power-plant/openapi-schema";
import type { Options as PowerPlantOpenAPITypeScriptOptions } from "@power-plant/openapi-typescript";
import type {
  PowerPlantPluginContext,
  PowerPlantPluginResolvedConfig
} from "@powerlines/plugin-power-plant/types/plugin";
import type { ResolvedConfig, UserConfig } from "powerlines";

export type { Options as PowerPlantOpenAPITypeScriptOptions } from "@power-plant/openapi-typescript";
export type { OpenAPISchema } from "@power-plant/openapi-schema";

/**
 * Options for the OpenAPI Powerlines plugin.
 */
export type OpenAPIPluginOptions = Partial<
  Omit<PowerPlantOpenAPITypeScriptOptions, "output">
> & {
  /**
   * The OpenAPI schema to generate from.
   *
   * @remarks
   * This can be an {@link OpenAPISchema} object, a local file path, a remote URL
   * string, or a {@link URL} object.
   *
   * @defaultValue "\{root\}/schema.yaml"
   */
  schema?: string | URL | OpenAPISchema;

  /**
   * The path to the output file generated by the plugin.
   *
   * @remarks
   * If using the default value, the file can be imported from "powerlines:api". This field allows the use of the "\{builtinsPath\}" token to refer to the built-in Powerlines plugins directory - the value will be replaced with the correct file path by the plugin.
   *
   * @defaultValue "\{builtinsPath\}/api.ts"
   */
  outputPath?: string;
};

export type OpenAPIPluginUserConfig = UserConfig & {
  openapi?: OpenAPIPluginOptions;
};

export type OpenAPIPluginResolvedConfig = ResolvedConfig &
  PowerPlantPluginResolvedConfig<
    OpenAPISchema,
    PowerPlantOpenAPITypeScriptOptions
  > & {
    openapi: Omit<OpenAPIPluginOptions, "schema" | "outputPath"> & {
      schema: string | URL | OpenAPISchema;
      outputPath: string;
      document: OpenAPISchema;
    };
  };

export type OpenAPIPluginContext<
  TResolvedConfig extends OpenAPIPluginResolvedConfig =
    OpenAPIPluginResolvedConfig
> = PowerPlantPluginContext<
  OpenAPISchema,
  PowerPlantOpenAPITypeScriptOptions,
  TResolvedConfig
>;